requirements.txt
djangorestframework
python-decouple
google-generativeai
spacy
scikit-learn